sclang: Refactor bytecodes
Rather than do all the bit shifting and magic numbers in the parse nodes classes, this moves the responsibility into dedicated opcode classes, making things easier to read.

This would not have been possible without Moss Heim's work documenting the byte codes. Their documentation is copied into the source code as comments above the respective opcodes so when you hover over the opcode in an IDE, their helpful descriptions appears in the pop-up.

revvert

+++ lang/LangSource/PyrParseNode.h
@@ -475,7 +475,7 @@ int nodeListLength(PyrParseNode* node);
 bool isSuperObjNode(PyrParseNode* node);
 bool isThisObjNode(PyrParseNode* node);
 int conjureSelectorIndex(PyrParseNode* node, PyrBlock* func, bool isSuper, PyrSymbol* selector, int* selType);
-int conjureLiteralSlotIndex(PyrParseNode* node, PyrBlock* func, PyrSlot* slot);
+Byte conjureLiteralSlotIndex(PyrParseNode* node, PyrBlock* func, PyrSlot* slot);
 bool findVarName(PyrBlock* func, PyrClass** classobj, PyrSymbol* name, int* varType, int* level, int* index,
                  PyrBlock** tempfunc);
 void countClassVarDefs(PyrClassNode* node, int* numClassMethods, int* numInstMethods);
